requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
requests==2.32.4
psycopg2-binary==2.9.10
gunicorn==23.0.0
djangorestframework-simplejwt

